+++ exec/Kathu/App/World.hs
@@ -5,7 +5,7 @@ import           Control.Monad                   (void)
 
 import           Kathu.App.Graphics.Image        (ImageID)
 import           Kathu.App.Graphics.ImageManager
-import           Kathu.App.System                (SystemT', destroyEntity, externalFunctions, newFromPrototype)
+import           Kathu.App.System
 import qualified Kathu.Scripting.Lua             as Lua
 import           Kathu.World.WorldSpace
 
@@ -14,7 +14,7 @@ loadWorldSpace ws = do
     prevManager <- get global
     manager <- loadPalettes (worldPalettes ws) prevManager
     global  $= manager
-    initWorldSpace destroyEntity newFromPrototype (Lua.loadScript externalFunctions) ws
+    initWorldSpace destroyEntity newFromPrototypeWithScriptMapping (Lua.loadScript id externalFunctions) ws
     void $ setPaletteManager (initialPalette ws)
 
     -- to prevent pausing issues during gameplay, we force a GC now while it's just done loading