+++ library/Kathu/Scripting/Lua.hs
@@ -13,6 +13,7 @@ module Kathu.Scripting.Lua
     , call
     , loadScript
     , shouldScriptRun
+    , setInstanceConfig
     , addWireController
     , addWireReceiver
     , releaseActiveScript
@@ -26,12 +27,14 @@ import           Control.Monad                     (unless, when)
 import           Control.Monad.IO.Class            (liftIO)
 import           Control.Monad.ST                  (stToIO)
 import qualified Data.HashTable.ST.Basic           as HT
+import qualified Data.Map                          as Map
 import qualified Data.Vector                       as Vec
 import           Foreign.Lua                       hiding (call, error, runWith)
 import qualified Foreign.Lua.Core                  as Lua
 import qualified Foreign.Lua.FunctionCalling       as Lua
 
 import           Kathu.Entity.Components
+import           Kathu.Entity.Logger
 import           Kathu.Entity.System
 import           Kathu.Entity.Time
 import           Kathu.Graphics.Camera
@@ -48,15 +51,14 @@ import           Kathu.Util.Types
 
 type HasScripting w m = ReadWriteEach w m [ActiveScript, RunningScriptEntity, ScriptBank, ScriptEventBuffer, WireReceivers]
 
--- in future an actual log file should be used and acessed through the world
-logLua :: String -> Lua ()
-logLua = liftIO . putStrLn
-
 call :: Lua.LuaCallFunc a => String -> a
 call = callFunc
 
 shouldScriptRun :: ScriptEvent -> ActiveScript -> Bool
 shouldScriptRun e = isEventSet e . eventFlags
+
+setInstanceConfig :: IDMap WorldVariable -> ActiveScript -> ActiveScript
+setInstanceConfig config script = script {instanceConfig = config}
 
 addWireController :: Identifier -> ActiveScript -> ActiveScript
 addWireController sigName as = as { wireSignals = Vec.cons sigName (wireSignals as) }
@@ -74,17 +76,24 @@ addWireReceiver sigName as@ActiveScript {instanceEntity = ety} = do
             
     liftIO $ addWireListener sigName (unEntity ety) onChangeIO receivers
 
-mkActiveScript :: forall w. (ReadWrite w IO ScriptBank) => Entity -> SingletonStatus -> Lua () -> Script -> SystemT w IO ActiveScript
-mkActiveScript !ety !singStatus !initLua (Script _ mainScr flags _) = mkActive <$> liftIO stateMVar
-    where mkActive st = ActiveScript st flags ety Vec.empty Vec.empty singStatus
-          !stateMVar = do
-              st  <- Lua.newstate
-              -- for some reason Lua exception logging doesn't actually get caught here, unlike in execFor...
-              st' <- Lua.runWith st (initLua >> handleLuaOp (Lua.dostring mainScr) >> Lua.state)
-              newMVar st'
+mkActiveScript :: forall w. (HasScripting w IO) => (ActiveScript -> ActiveScript) -> Entity -> SingletonStatus -> Lua () -> Script -> SystemT w IO ActiveScript
+mkActiveScript mapper !ety !singStatus !initLua (Script _ mainScr flags _) = do
+    mvar <- liftIO newEmptyMVar
+    -- warning: by this point no mvar is supplied, so if something tries to use it it will have issues
+    -- fortunately nothing should, as this won't trigger execFor/runFor or events
+    let baseAS = mapper $ ActiveScript mvar flags ety Vec.empty Vec.empty singStatus Map.empty
+    ety    $= baseAS
+    global $= RunningScriptEntity (Just ety)
+
+    mvar' <- liftIO $ do
+        st  <- Lua.newstate
+        -- for some reason Lua exception logging doesn't actually get caught here, unlike in execFor...
+        st' <- Lua.runWith st (initLua >> handleLuaOp (Lua.dostring mainScr) >> Lua.state)
+        newMVar st'
+    pure $ baseAS {activeState = mvar'}
 
 releaseActiveScript :: forall w. (ReadWriteEach w IO [RunningScriptEntity, ScriptEventBuffer, Variables, WireReceivers]) => ActiveScript -> SystemT w IO ()
-releaseActiveScript as@(ActiveScript stmvar _ scriptEntity watched signals singStatus) = do
+releaseActiveScript as@(ActiveScript stmvar _ scriptEntity watched signals singStatus _) = do
     let ety = unEntity scriptEntity
     when (scriptEntity /= global && shouldScriptRun onDestroy as) $
         execFor as (call "onDestroy" ety)
@@ -103,19 +112,18 @@ releaseActiveScript as@(ActiveScript stmvar _ scriptEntity watched signals singS
             Lua.close lstate
             putMVar stmvar $ error "Attempted to use a release ActiveScript"
 
-loadScript :: forall w g. (Has w IO Physics, Members w IO (Render g), ReadWriteEach w IO [ActiveScript, Camera, Debug, Force, Identity, Local, LogicTime, Mass, MovingSpeed, Position, Random, Render g, RenderTime, RunningScriptEntity, ScriptBank, ScriptEventBuffer, Tags, Variables, Velocity, WireReceivers])
-           => ExternalFunctions w g -> Entity -> Script -> SystemT w IO ActiveScript
-loadScript extFuns ety script
+loadScript :: forall w g. (Has w IO Physics, Members w IO (Render g), ReadWriteEach w IO [ActiveScript, Camera, Debug, Force, Identity, Local, Logger, LogicTime, Mass, MovingSpeed, Position, Random, Render g, RenderTime, RunningScriptEntity, ScriptBank, ScriptEventBuffer, Tags, Variables, Velocity, WireReceivers])
+           => (ActiveScript -> ActiveScript) -> ExternalFunctions w g -> Entity -> Script -> SystemT w IO ActiveScript
+loadScript mapper extFuns ety script
     | isSingleton script = runIfOnInit =<< fromBank
-    | otherwise          = runIfOnInit =<< mkAS NonSingleton ety
-    where mkAS singStatus e = ask >>= (\l -> mkActiveScript e singStatus l script) . initLua
+    | otherwise          = runIfOnInit =<< mkAS mapper NonSingleton ety
+    where mkAS f singStatus e = ask >>= (\l -> mkActiveScript f e singStatus l script) . initLua
           initLua world = do
               openbase
               openmath
               openpackage
               openstring
               opentable
-              registerHaskellFunction "log" logLua
               registerComponentFunctions world extFuns
               registerGlobalFunctions world extFuns
           runIfOnInit as | shouldScriptRun onInit as = set ety as >> execFor as (call "onInit" (unEntity ety)) >> return as
@@ -125,10 +133,10 @@ loadScript extFuns ety script
               sbank     <- unScriptBank <$> get global
               curScript <- liftIO . stToIO $ HT.lookup sbank sID
               case curScript of
-                  Just as -> return $ as {instanceEntity = ety, singletonStatus = SingletonReference}
-                  Nothing -> do ascript <- mkAS SingletonBase global
+                  Just as -> pure . mapper $ as {instanceEntity = ety, singletonStatus = SingletonReference}
+                  Nothing -> do ascript <- mkAS id SingletonBase global
                                 liftIO . stToIO $ HT.insert sbank sID ascript
-                                return $ ascript {instanceEntity = ety, singletonStatus = SingletonReference}
+                                pure . mapper $ ascript {instanceEntity = ety, singletonStatus = SingletonReference}
 
 initScripting :: forall w. (ReadWriteEach w IO [ScriptBank, WireReceivers]) => SystemT w IO ()
 initScripting = do