+++ library/Kathu/Scripting/Lua/Component.hs
@@ -1,46 +1,56 @@
 {-# LANGUAGE BangPatterns        #-}
 {-# LANGUAGE DataKinds           #-}
 {-# LANGUAGE FlexibleContexts    #-}
+{-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Kathu.Scripting.Lua.Component (registerComponentFunctions) where
 
 import           Apecs
 import           Apecs.Physics
-import           Control.Monad                     (forM_)
-import           Data.Text                         (Text)
-import qualified Data.Vector                       as Vec
+import           Control.Monad                       (forM_, when)
+import qualified Data.Map                            as Map
+import           Data.Maybe                          (fromJust, isJust, isNothing)
+import           Data.Text                           (Text)
+import qualified Data.Vector                         as Vec
+import qualified Data.Set                            as DSet
+import qualified Data.Text                           as T
 import           Foreign.Lua
-import           Linear.V2                         (V2(..))
+import           Linear.V2                           (V2(..))
 
 import           Kathu.Entity.Components
-import           Kathu.Graphics.Drawable           (Render(..), RenderSprite(..), switchAnimationByID)
+import           Kathu.Entity.Logger
+import           Kathu.Entity.Physics.CollisionGroup
+import           Kathu.Graphics.Drawable             (Render(..), RenderSprite(..), switchAnimationByID)
 import           Kathu.Scripting.ExternalFunctions
 import           Kathu.Scripting.Lua.Types
+import           Kathu.Scripting.Variables
 import           Kathu.Scripting.Wire
 import           Kathu.Util.Apecs
-import           Kathu.Util.Types                  (mkIdentifier, unID)
+import           Kathu.Util.Types
 
 -- ExternalFunctions is unused in this, but is included here since it might be in the future, mirrors the Global's function signature, and acts as a Proxy for g
-registerComponentFunctions :: forall w g. (Has w IO Physics, Members w IO (Render g), ReadWriteEach w IO '[ActiveScript, Force, Identity, Mass, MovingSpeed, Position, Render g, RunningScriptEntity, ScriptEventBuffer, Tags, Velocity, WireReceivers])
+registerComponentFunctions :: forall w g. (Has w IO Physics, Members w IO (Render g), ReadWriteEach w IO '[ActiveScript, Force, Identity, Logger, Mass, MovingSpeed, Position, Render g, RunningScriptEntity, ScriptEventBuffer, Tags, Velocity, WireReceivers])
                            => w -> ExternalFunctions w g -> Lua ()
 registerComponentFunctions world _ = do
-    registerHaskellFunction "getIdentifier"   $ getIdentifier world
-    registerHaskellFunction "getName"         $ getName world
-    registerHaskellFunction "getDescription"  $ getDescription world
-    registerHaskellFunction "getTags"         $ getTags world
-    registerHaskellFunction "getMovingSpeed"  $ getMovingSpeed world
-    registerHaskellFunction "setMovingSpeed"  $ setMovingSpeed world
-    registerHaskellFunction "setAnimation"    $ setAnimation (Proxy :: Proxy g) world
-    registerHaskellFunction "getMass"         $ getMass world
-    registerHaskellFunction "setMass"         $ setMass world
-    registerHaskellFunction "getPosition"     $ getPosition world
-    registerHaskellFunction "setPosition"     $ setPosition world
-    registerHaskellFunction "getVelocity"     $ getVelocity world
-    registerHaskellFunction "setVelocity"     $ setVelocity world
-    registerHaskellFunction "getForce"        $ getForce world
-    registerHaskellFunction "setForce"        $ setForce world
-    registerHaskellFunction "modifyWirePower" $ modifyWirePower world
+    registerHaskellFunction "getIdentifier"        $ getIdentifier world
+    registerHaskellFunction "getName"              $ getName world
+    registerHaskellFunction "getDescription"       $ getDescription world
+    registerHaskellFunction "getTags"              $ getTags world
+    registerHaskellFunction "getMovingSpeed"       $ getMovingSpeed world
+    registerHaskellFunction "setMovingSpeed"       $ setMovingSpeed world
+    registerHaskellFunction "setAnimation"         $ setAnimation (Proxy :: Proxy g) world
+    registerHaskellFunction "getMass"              $ getMass world
+    registerHaskellFunction "setMass"              $ setMass world
+    registerHaskellFunction "getPosition"          $ getPosition world
+    registerHaskellFunction "setPosition"          $ setPosition world
+    registerHaskellFunction "getVelocity"          $ getVelocity world
+    registerHaskellFunction "setVelocity"          $ setVelocity world
+    registerHaskellFunction "getForce"             $ getForce world
+    registerHaskellFunction "setForce"             $ setForce world
+    registerHaskellFunction "setCollisionCategory" $ setCollisionCategory world
+    registerHaskellFunction "modifyWirePower"      $ modifyWirePower world
+    registerHaskellFunction "getConfig"            $ getInstanceConfigVariable world
 
 getIdentifier :: forall w. (ReadWrite w IO Identity) => w -> Int -> Lua (Optional Text)
 getIdentifier !world !etyID = liftIO . Apecs.runWith world $ do
@@ -67,7 +77,7 @@ getTags :: forall w. (ReadWrite w IO Tags) => w -> Int -> Lua (Optional [Text])
 getTags !world !etyID = liftIO . Apecs.runWith world $ do
     tags <- getIfExists (Entity etyID)
     pure $ case tags of
-        Just (Tags t) -> Optional $ Just (Vec.toList t)
+        Just (Tags t) -> Optional $ Just (DSet.toList t)
         Nothing       -> Optional Nothing
 
 getMovingSpeed :: forall w. (ReadWrite w IO MovingSpeed) => w -> Int -> Lua (Optional Double)
@@ -96,15 +106,33 @@ setAnimation _ !world !etyID !animID = liftIO . Apecs.runWith world $ do
         Nothing     -> return ()
         Just (Render layers) -> ety $= Render (changeAnim <$> layers)
 
+setCollisionCategory :: forall w. (Has w IO Physics, ReadWriteEach w IO [CollisionFilter, Tags]) => w -> Entity -> Text -> Optional Text -> Lua ()
+setCollisionCategory !world !ety colCategory (Optional tag) = liftIO . Apecs.runWith world $ do
+    ShapeList shapes <- get ety
+    
+    let maybeColGroup = collisionGroupFromString colCategory
+    case groupCollisionFilter <$> maybeColGroup of
+        Nothing        -> lift . putStrLn $ "Couldn't find collision category " ++ show colCategory
+        Just colFilter ->
+            forM_ shapes $ \sEty -> do
+                maybeTags <- get sEty
+                let colGroup = fromJust maybeColGroup
+                case maybeTags of
+                    Nothing          ->
+                        sEty $= (colFilter, mkGroupSensor colGroup)
+                    Just (Tags tags) ->
+                        when (isJust tag && DSet.member (fromJust tag) tags) $
+                            sEty $= (colFilter, mkGroupSensor colGroup)
+
 -------------
 -- Physics --
 -------------
 
-getVector2D :: forall w c. (Get w IO c, Has w IO c, Members w IO c) => (c -> V2 Double) -> w -> Int -> Lua (Optional (Double, Double))
+getVector2D :: forall w c. (Get w IO c, Has w IO c, Members w IO c) => (c -> V2 Double) -> w -> Int -> Lua (Optional (V2 Double))
 getVector2D mapper !world !etyID = liftIO . Apecs.runWith world $ do
     comp <- getIfExists (Entity etyID)
     pure $ case comp of
-        Just vec  -> let (V2 x y) = mapper vec in Optional $ Just (x, y)
+        Just vec  -> Optional . Just . mapper $ vec
         Nothing   -> Optional Nothing
 
 getMass :: forall w. (ReadWrite w IO Mass) => w -> Int -> Lua (Optional Double)
@@ -114,30 +142,30 @@ getMass !world !etyID = liftIO . Apecs.runWith world $ do
         Just (Mass m) -> Optional $ Just m
         Nothing       -> Optional Nothing
 
-setMass :: forall w. (Has w IO Physics, ReadWrite w IO Position) => w -> Int -> Double-> Lua ()
+setMass :: forall w. (Has w IO Physics) => w -> Int -> Double-> Lua ()
 setMass !world !etyID m = liftIO . Apecs.runWith world $
     Entity etyID $= Mass m
 
-getPosition :: forall w. (ReadWrite w IO Position) => w -> Int -> Lua (Optional (Double, Double))
+getPosition :: forall w. (ReadWrite w IO Position) => w -> Int -> Lua (Optional (V2 Double))
 getPosition = getVector2D $ \(Position v) -> v
 
-setPosition :: forall w. (ReadWrite w IO Position) => w -> Int -> (Double, Double) -> Lua ()
-setPosition !world !etyID (x, y) = liftIO . Apecs.runWith world $
-    Entity etyID $= Position (V2 x y)
+setPosition :: forall w. (ReadWrite w IO Position) => w -> Int -> V2 Double -> Lua ()
+setPosition !world !etyID v = liftIO . Apecs.runWith world $
+    Entity etyID $= Position v
 
-getVelocity :: forall w. (ReadWrite w IO Velocity) => w -> Int -> Lua (Optional (Double, Double))
+getVelocity :: forall w. (ReadWrite w IO Velocity) => w -> Int -> Lua (Optional (V2 Double))
 getVelocity = getVector2D $ \(Velocity v) -> v
 
-setVelocity :: forall w. (ReadWrite w IO Velocity) => w -> Int -> (Double, Double) -> Lua ()
-setVelocity !world !etyID (x, y) = liftIO . Apecs.runWith world $
-    Entity etyID $= Velocity (V2 x y)
+setVelocity :: forall w. (ReadWrite w IO Velocity) => w -> Int -> V2 Double -> Lua ()
+setVelocity !world !etyID v = liftIO . Apecs.runWith world $
+    Entity etyID $= Velocity v
 
-getForce :: forall w. (ReadWrite w IO Force) => w -> Int -> Lua (Optional (Double, Double))
+getForce :: forall w. (ReadWrite w IO Force) => w -> Int -> Lua (Optional (V2 Double))
 getForce = getVector2D $ \(Force v) -> v
 
-setForce :: forall w. (ReadWrite w IO Force) => w -> Int -> (Double, Double) -> Lua ()
-setForce !world !etyID (x, y) = liftIO . Apecs.runWith world $
-    Entity etyID $= Force (V2 x y)
+setForce :: forall w. (ReadWrite w IO Force) => w -> Int -> V2 Double -> Lua ()
+setForce !world !etyID v = liftIO . Apecs.runWith world $
+    Entity etyID $= Force v
 
 -- Wires --
 
@@ -148,3 +176,18 @@ modifyWirePower !world !etyID !dPower = liftIO . Apecs.runWith world $ do
 
     forM_ maybeScript $ \script -> Vec.forM_ (wireSignals script) $ \signalID ->
         mutateWirePower signalID (+dPower) receivers
+
+-- Script --
+
+getInstanceConfigVariable :: forall w. (ReadWriteEach w IO [ActiveScript, Logger, RunningScriptEntity]) => w -> Identifier -> Lua (Optional WorldVariable)
+getInstanceConfigVariable !world !idt = liftIO . Apecs.runWith world $ do
+    maybeEty <- runningScript <$> get global
+    case maybeEty of
+        Nothing  -> pure $ Optional Nothing
+        Just ety -> do
+            script    <- get ety
+            let config = instanceConfig script
+                var    = Map.lookup idt config
+            when (isNothing var) $ 
+                logLine Warning $ T.concat ["Config variable ", T.pack (show idt),  " is not present in the instance config for entity ", T.pack (show ety), " with config ", T.pack (show config), ", was this function called during the initial script parsing?"]
+            pure . Optional $ var